QUANTUMIDE/quantum-ide_requirements.txt
flask==3.0.3
flask-cors==4.0.0
pyyaml==6.0.1
qutip==5.0.3
numpy==1.26.4